Fix authorized_keys permissions
It turns out that the aforementioned file will end up with root being the owner and group of that file. This will prevent the ansible user from being able to login using the key. And, also, it will be a very difficult bug to catch. 

This is fixed in this patch.