fix exercise131.yaml
It seems that using a template to loop through several groups in order to add sudo rights does not generate the desired state.  We end up overwriting the file and, in my case, the only entry allows sudo for the students group only. 

If we want all groups to be added to sudo, we need to use lineinfile.